move step groups into its own file
